third_party/blink/web_tests/external/wpt/editing/run/insertimage-expected.txt

This is a testharness.js-based test.
Found 18 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertimage",""]] "foo[bar]baz": execCommand("insertimage", false, "") return value
  assert_equals: expected false but got true
[FAIL] [["insertimage",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobarbaz" but got "foo<img>baz"
[FAIL] [["stylewithcss","true"],["insertimage","/img/lion.svg"]] "[foo<span style=color:#aBcDeF>bar]</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo</p>{<p>bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><img src=\\"/img/lion.svg\\"><p>baz</p>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\">baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo</p>{<p>bar</p>}<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><img src=\\"/img/lion.svg\\"><p>baz</p>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\">baz</p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\"><font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<img src=\\"/img/lion.svg\\"><span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["stylewithcss","false"],["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo[bar<p style=color:blue>baz]quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\"><font color=\\"#0000ff\\">quz</font></p>" but got "<p>foo<img src=\\"/img/lion.svg\\"><span style=\\"color:rgb(0, 0, 255)\\">quz</span></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><img src=\\"/img/lion.svg\\"></p></div>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\"></p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<div><p>foo<p>[bar<p>baz]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p><p><img src=\\"/img/lion.svg\\"></p></div>" but got "<p>foo</p><p><img src=\\"/img/lion.svg\\"></p>"
[FAIL] [["defaultparagraphseparator","div"],["insertimage","/img/lion.svg"]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar<br>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>"
[FAIL] [["defaultparagraphseparator","p"],["insertimage","/img/lion.svg"]] "<p>foo[</p><p>]bar<br>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar<br>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p>" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar<br>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p>baz" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar<br>baz"
[FAIL] [["insertimage","/img/lion.svg"]] "<p>foo[</p>]bar<p>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<img src=\\"/img/lion.svg\\">bar</p><p>baz</p>" but got "<p>foo<img src=\\"/img/lion.svg\\"></p>bar<p>baz</p>"
[FAIL] [["insertimage","/img/lion.svg"]] "<div><p>foo</p>bar[</div>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><p>foo</p>bar<img src=\\"/img/lion.svg\\">baz</div>" but got "<div><p>foo</p>bar<img src=\\"/img/lion.svg\\"></div>baz"
[FAIL] [["insertimage","/img/lion.svg"]] "<div><b>{}<br></b></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><img src=\\"/img/lion.svg\\"><b></b></div>" not in array ["<div><img src=\\"/img/lion.svg\\"></div>", "<div><img src=\\"/img/lion.svg\\"><b><br></b></div>"]
[FAIL] [["insertimage","/img/lion.svg"]] "<div><span>{}<br></span></div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div><img src=\\"/img/lion.svg\\"><span></span></div>" not in array ["<div><img src=\\"/img/lion.svg\\"></div>", "<div><img src=\\"/img/lion.svg\\"><span><br></span></div>"]
[FAIL] [["insertimage","/img/lion.svg"]] "<div><span>A[]B</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span>A</span><img src=\\"/img/lion.svg\\"><span>B</span></div>" but got "<div><span>A</span><img src=\\"/img/lion.svg\\">B</div>"
Harness: the test ran to completion.
